backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
PyJWT[crypto]==2.10.1
httpx==0.27.2
supabase==2.10.0
APScheduler==3.11.0
pandas==2.2.3
scikit-learn==1.6.0
python-socketio==5.12.0
slowapi==0.1.9
python-dotenv==1.0.1
loguru==0.7.3
Pillow==11.0.0
